:recycle: Refactor: Declare namespace to patch modules.
